Remove homeopathic tracking of interaction event like "click" or "focus".
We only waited until the end of this task before stopping the work. There is no way Capybara can observe the JavaScript state before the end of this task anyway.

If e.g. a click event handler starts an AJAX request, its work will be tracked before the end of this task.