grammatical tweaks in README.md
I noticed a few grammatical issues in the early part of the README, so I went through about the first 150 lines and made as many improvements as I could.

But at line 147 I ran into an issue I can't fix without asking a question.  I think maybe a word is missing, and I can't guess what it is.

> Using the middleware is optional, as the JavaScript snippet already for AJAX requests on the client.

What does the JavaScript snippet already do?

If you agree that these are improvements and would like me to continue on the rest of the README, just let me know.  And if you can tell me what the sentence above was supposed to say, I'll do my best to make it as clear as possible using standard English syntax.